python-ocr/requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0


















